fix: shan's wish is my command
