+++ deno.json
@@ -1,10 +1,14 @@
 {
+  "imports": {
+    "@discordeno": "https://deno.land/x/discordeno@18.0.1/mod.ts",
+    "@irc/client": "jsr:@irc/client@0.17.2",
+    "lru-cache": "npm:lru-cache@11.0.0"
+  },
   "tasks": {
-    "run": "deno run --allow-net --allow-read --allow-import src/app/main.ts",
-    "fmt": "deno fmt src/app src/bridge src/testing AGENTS.md README.md docs/ARCHITECTURE.md CHANGELOG.md",
-    "lint": "deno lint src/app src/bridge src/testing",
-    "check": "deno check --allow-import src/app/main.ts src/bridge/index.ts",
-    "test": "deno test --allow-import src/testing",
-    "validate": "deno fmt --check src/app src/bridge src/testing AGENTS.md README.md docs/ARCHITECTURE.md CHANGELOG.md && deno lint src/app src/bridge src/testing && deno check --allow-import src/app/main.ts src/bridge/index.ts && deno test --allow-import src/testing"
+    "run": "deno run --allow-net --allow-read --allow-import src/main.ts",
+    "fmt": "deno fmt src README.md CHANGELOG.md",
+    "lint": "deno lint src",
+    "check": "deno check --allow-import src/main.ts src/bridge/index.ts",
+    "validate": "deno fmt --check src README.md CHANGELOG.md && deno lint src && deno check --allow-import src/main.ts src/bridge/index.ts"
   }
 }
